test(unit): convert tests to vitest
