+++ .eslintrc.yml
@@ -1,3 +1,2 @@
 extends:
   - '@form8ion'
-  - '@form8ion/mocha'
