--- .mocharc.json
@@ -1,1 +1,0 @@
-{"ui":"tdd","require":["@babel/register","./test/mocha-setup.js"],"recursive":true}
