+++ package.json
@@ -28,7 +28,7 @@
     "watch": "run-s 'build:js -- --watch'",
     "prepack": "run-s build",
     "test:unit": "cross-env NODE_ENV=test c8 run-s test:unit:base",
-    "test:unit:base": "DEBUG=any mocha 'src/**/*-test.js'",
+    "test:unit:base": "DEBUG=any vitest run",
     "prepare": "husky install",
     "lint:publish": "publint --strict",
     "lint:js:fix": "run-s 'lint:js -- --fix'",
@@ -44,32 +44,29 @@
   },
   "packageManager": "npm@11.2.0+sha512.3dc9c50ba813a3d54393155a435fe66404b72685ab0e3008f9ae9ed8d81f6104860f07ed2656dd5748c1322d95f3140fa9b19c59a6bba7750fd12285f81866da",
   "dependencies": {
-    "@form8ion/core": "^4.0.0",
-    "make-dir": "^4.0.0"
+    "@form8ion/core": "^4.0.0"
   },
   "devDependencies": {
     "@babel/register": "7.25.9",
     "@form8ion/babel-preset": "1.6.143",
     "@form8ion/commitlint-config": "2.0.5",
-    "@form8ion/eslint-config": "6.0.5",
-    "@form8ion/eslint-config-mocha": "3.0.6",
+    "@form8ion/eslint-config": "7.0.12",
     "@form8ion/remark-lint-preset": "6.0.7",
     "@travi/any": "3.1.2",
     "ban-sensitive-files": "1.10.7",
     "c8": "10.1.3",
-    "chai": "5.2.0",
     "cross-env": "7.0.3",
     "cz-conventional-changelog": "3.3.0",
     "husky": "9.1.7",
     "lockfile-lint": "4.14.0",
     "ls-engines": "0.9.3",
-    "mocha": "11.1.0",
     "npm-run-all2": "7.0.2",
     "publint": "0.3.9",
     "remark-cli": "12.0.1",
     "rimraf": "6.0.1",
     "rollup": "4.35.0",
     "rollup-plugin-auto-external": "2.0.0",
-    "sinon": "19.0.2"
+    "vitest": "3.0.8",
+    "vitest-when": "0.6.0"
   }
 }
