+++ src/scaffolder.js
@@ -1,12 +1,10 @@
-import {resolve} from 'path';
-import {promises} from 'fs';
+import {resolve} from 'node:path';
+import {promises as fs} from 'node:fs';
 import {fileTypes, writeConfigFile} from '@form8ion/core';
 
-import mkdir from '../thirdparty-wrappers/make-dir';
-
 export async function scaffold({projectRoot, testDirectory, testBaseUrl}) {
-  const [createdTestDirectory] = await Promise.all([
-    mkdir(testDirectory),
+  await Promise.all([
+    fs.mkdir(testDirectory),
     writeConfigFile({
       path: projectRoot,
       name: 'cypress',
@@ -15,9 +13,9 @@ export async function scaffold({projectRoot, testDirectory, testBaseUrl}) {
     })
   ]);
 
-  await promises.copyFile(
+  await fs.copyFile(
     resolve(__dirname, '..', 'templates', 'canary-spec.js'),
-    `${createdTestDirectory}/canary-spec.js`
+    `${testDirectory}/canary-spec.js`
   );
 
   return {
