--- test/mocha-setup.js
@@ -1,8 +1,0 @@
-import sinon from 'sinon';
-import chai from 'chai';
-
-sinon.assert.expose(chai.assert, {prefix: ''});
-
-process.on('unhandledRejection', reason => {
-  throw reason;
-});
