--- thirdparty-wrappers/make-dir.js
@@ -1,3 +1,0 @@
-import makeDir from 'make-dir';
-
-export default makeDir;
