+++ vitest.config.js
@@ -1,0 +1,8 @@
+// eslint-disable-next-line import/no-unresolved
+import {defineConfig} from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    restoreMocks: true
+  }
+});
